[fix] meet: move page/API tests out of src/pages
