+++ apps/meet/src/__tests__/pages/api/status.test.ts
@@ -2,7 +2,7 @@
 import { type NextApiRequest, type NextApiResponse } from 'next';
 import { createMocks } from 'node-mocks-http';
 import { describe, expect, test } from 'vitest';
-import handle from './status';
+import handle from '../../../pages/api/status';
 
 describe('/api/status', () => {
   test('returns an online status', async () => {
